requirements.txt
feedparser>=6.0.10
requests>=2.31.0
google-generativeai>=0.8.0
flask>=3.0.0
gunicorn>=21.0.0
